+++ CHANGELOG.md
@@ -1,8 +1,11 @@
 # Changelog
 
-## Version 0.1.1
+## Version 0.2.0
 
-- Compatibility with FoundryVTT V14
+- Added a new configuration menu to set an actor’s daily food and water limits, as well as the number of days before they begin to starve. The menu is available in actor sheet edit mode. (#1)
+- Added support for metric units when the metric setting is enabled.
+- Included various smaller fixes and improvements.
+- Added compatibility with Foundry VTT V14.
 
 ## Version 0.1.0